+++ tbot/machine/channel/__init__.py
@@ -1,4 +1,4 @@
-from .channel import Channel, ChannelClosedException, SkipStream
+from .channel import Channel, ChannelClosedException, SkipStream, TBOT_PROMPT
 from .paramiko import ParamikoChannel
 from .subprocess import SubprocessChannel
 
@@ -8,4 +8,5 @@ __all__ = (
     "SkipStream",
     "ParamikoChannel",
     "SubprocessChannel",
+    "TBOT_PROMPT",
 )